float and friends list fix

+++ src/main/java/ir/ac/aut/hesabgar_group/domain/data/PaymentTerm.java
@@ -2,9 +2,9 @@ package ir.ac.aut.hesabgar_group.domain.data;
 
 public class PaymentTerm {
     private String user;
-    private int amount;
+    private float amount;
 
-    public PaymentTerm(String user, int amount) {
+    public PaymentTerm(String user, float amount) {
         this.user = user;
         this.amount = amount;
     }
@@ -17,11 +17,11 @@ public class PaymentTerm {
         this.user = user;
     }
 
-    public int getAmount() {
+    public float getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
 }
