+++ src/main/java/ir/ac/aut/hesabgar_group/request/AddingInvoiceRequest.java
@@ -7,15 +7,15 @@ public class AddingInvoiceRequest {
     private String groupId;
     private String topic;
     private String description;
-    private int totalPaidValue;
-    private Map<String, Integer> groupShare;
+    private float totalPaidValue;
+    private Map<String, Float> groupShare;
 
 
-    public int getTotalPaidValue() {
+    public float getTotalPaidValue() {
         return totalPaidValue;
     }
 
-    public void setTotalPaidValue(int totalPaidValue) {
+    public void setTotalPaidValue(float totalPaidValue) {
         this.totalPaidValue = totalPaidValue;
     }
 
@@ -51,11 +51,11 @@ public class AddingInvoiceRequest {
         this.description = description;
     }
 
-    public Map<String, Integer> getGroupShare() {
+    public Map<String, Float> getGroupShare() {
         return groupShare;
     }
 
-    public void setGroupShare(Map<String, Integer> groupShare) {
+    public void setGroupShare(Map<String, Float> groupShare) {
         this.groupShare = groupShare;
     }
